refactor : MenuValidator 구조 수정 및 테스트코드 통합
