+++ src/test/java/kitchenpos/order/application/TableGroupServiceTest.java
@@ -15,6 +15,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuValidator;
 import kitchenpos.menu.dto.application.MenuProductDto;
 import kitchenpos.menu.repository.MenuGroupRepository;
 import kitchenpos.menu.repository.MenuRepository;
@@ -49,6 +50,9 @@ class TableGroupServiceTest {
 
     @Autowired
     private MenuRepository menuRepository;
+
+    @Autowired
+    private MenuValidator menuValidator;
 
     @Nested
     @DisplayName("create()")
@@ -167,7 +171,8 @@ class TableGroupServiceTest {
             menuGroup.getId(),
             new ArrayList<MenuProductDto>() {{
                 add(new MenuProductDto(product.getId(), 1L));
-            }}
+            }},
+            menuValidator
         ));
 
         return new Order(
